Switch to using new config framework
