--- server/src/config/macros.rs
@@ -1,59 +1,0 @@
-/*
- * Created on Sat Oct 02 2021
- *
- * This file is a part of Skytable
- * Skytable (formerly known as TerrabaseDB or Skybase) is a free and open-source
- * NoSQL database written by Sayan Nandan ("the Author") with the
- * vision to provide flexibility in data modelling without compromising
- * on performance, queryability or scalability.
- *
- * Copyright (c) 2021, Sayan Nandan <ohsayan@outlook.com>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <https://www.gnu.org/licenses/>.
- *
-*/
-
-macro_rules! cli_parse_or_default_or_err {
-    ($parsewhat:expr, $default:expr, $except:expr $(,)?) => {
-        match $parsewhat.map(|v| v.parse()) {
-            Some(Ok(v)) => v,
-            Some(Err(_)) => return Err(self::cfgerr::ConfigError::CliArgErr($except)),
-            None => $default,
-        }
-    };
-}
-
-macro_rules! cli_setparse_or_err {
-    ($setwhat:expr, $parsewhat:expr, $except:expr $(,)?) => {
-        match $parsewhat.map(|v| v.parse()) {
-            Some(Ok(v)) => $setwhat = v,
-            Some(Err(_)) => return Err(self::cfgerr::ConfigError::CliArgErr($except)),
-            _ => {}
-        }
-    };
-}
-
-macro_rules! set_if_exists {
-    ($testwhat:expr, $trywhat:expr) => {
-        if let Some(testwhat) = $testwhat {
-            $trywhat = testwhat;
-        }
-    };
-}
-
-macro_rules! ret_cli_err {
-    ($what:expr) => {
-        return Err(self::cfgerr::ConfigError::CliArgErr($what))
-    };
-}
