+++ server/src/config/tests.rs
@@ -24,27 +24,75 @@
  *
 */
 
-use super::cfgerr::{ConfigError, ERR_CONFLICT};
-use super::{
-    BGSave, ConfigurationSet, IpAddr, PortConfig, SnapshotConfig, SnapshotPref, SslOpts,
-    DEFAULT_IPV4, DEFAULT_PORT, MAXIMUM_CONNECTION_LIMIT,
-};
-use clap::{load_yaml, App};
+use super::{Configset, PortConfig, DEFAULT_IPV4};
 pub(super) use libsky::TResult;
 use std::fs;
-use std::net::Ipv6Addr;
 
 #[test]
-fn test_config_toml_okayport() {
-    let file = r#"
-    [server]
-    host = "127.0.0.1"
-    port = 2003
-"#
-    .to_owned();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(cfg, ConfigurationSet::default(),);
+fn server_tcp() {
+    let mut cfgset = Configset::new_env();
+    cfgset.server_tcp(
+        Some("127.0.0.1"),
+        "SKY_SERVER_HOST",
+        Some("2004"),
+        "SKY_SERVER_PORT",
+    );
+    assert_eq!(
+        cfgset.cfg.ports,
+        PortConfig::new_insecure_only(DEFAULT_IPV4, 2004)
+    );
+    assert!(cfgset.is_mutated());
+    assert!(cfgset.is_okay());
 }
+#[test]
+fn server_tcp_fail_host() {
+    let mut cfgset = Configset::new_env();
+    cfgset.server_tcp(
+        Some("?127.0.0.1"),
+        "SKY_SERVER_HOST",
+        Some("2004"),
+        "SKY_SERVER_PORT",
+    );
+    assert_eq!(
+        cfgset.cfg.ports,
+        PortConfig::new_insecure_only(DEFAULT_IPV4, 2004)
+    );
+    assert!(cfgset.is_mutated());
+    assert!(!cfgset.is_okay());
+}
+#[test]
+fn server_tcp_fail_port() {
+    let mut cfgset = Configset::new_env();
+    cfgset.server_tcp(
+        Some("127.0.0.1"),
+        "SKY_SERVER_HOST",
+        Some("65537"),
+        "SKY_SERVER_PORT",
+    );
+    assert_eq!(
+        cfgset.cfg.ports,
+        PortConfig::new_insecure_only(DEFAULT_IPV4, 2003)
+    );
+    assert!(cfgset.is_mutated());
+    assert!(!cfgset.is_okay());
+}
+#[test]
+fn server_tcp_fail_both() {
+    let mut cfgset = Configset::new_env();
+    cfgset.server_tcp(
+        Some("?127.0.0.1"),
+        "SKY_SERVER_HOST",
+        Some("65537"),
+        "SKY_SERVER_PORT",
+    );
+    assert_eq!(
+        cfgset.cfg.ports,
+        PortConfig::new_insecure_only(DEFAULT_IPV4, 2003)
+    );
+    assert!(cfgset.is_mutated());
+    assert!(!cfgset.is_okay());
+}
+
 /// Gets a `toml` file from `WORKSPACEROOT/examples/config-files`
 fn get_toml_from_examples_dir(filename: String) -> TResult<String> {
     use std::path;
@@ -55,217 +103,4 @@ fn get_toml_from_examples_dir(filename: String) -> TResult<String> {
     fileloc.push("config-files");
     fileloc.push(filename);
     Ok(fs::read_to_string(fileloc)?)
-}
-
-#[test]
-fn test_config_toml_badport() {
-    let file = r#"
-    [server]
-    port = 20033002
-"#
-    .to_owned();
-    let cfg = ConfigurationSet::new_from_toml_str(file);
-    assert!(cfg.is_err());
-}
-
-#[test]
-fn test_config_file_ok() {
-    let file = get_toml_from_examples_dir("skyd.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(cfg, ConfigurationSet::default());
-}
-
-#[test]
-fn test_config_file_err() {
-    let file = get_toml_from_examples_dir("skyd.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_file(file);
-    assert!(cfg.is_err());
-}
-#[test]
-fn test_args() {
-    let cmdlineargs = vec!["skyd", "--withconfig", "../examples/config-files/skyd.toml"];
-    let cfg_layout = load_yaml!("../cli.yml");
-    let matches = App::from_yaml(cfg_layout).get_matches_from(cmdlineargs);
-    let filename = matches.value_of("config").unwrap();
-    assert_eq!("../examples/config-files/skyd.toml", filename);
-    let cfg =
-        ConfigurationSet::new_from_toml_str(std::fs::read_to_string(filename).unwrap()).unwrap();
-    assert_eq!(cfg, ConfigurationSet::default());
-}
-
-#[test]
-fn test_config_file_noart() {
-    let file = get_toml_from_examples_dir("secure-noart.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(
-        cfg,
-        ConfigurationSet {
-            noart: true,
-            bgsave: BGSave::default(),
-            snapshot: SnapshotConfig::default(),
-            ports: PortConfig::default(),
-            maxcon: MAXIMUM_CONNECTION_LIMIT
-        }
-    );
-}
-
-#[test]
-fn test_config_file_ipv6() {
-    let file = get_toml_from_examples_dir("ipv6.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(
-        cfg,
-        ConfigurationSet {
-            noart: false,
-            bgsave: BGSave::default(),
-            snapshot: SnapshotConfig::default(),
-            ports: PortConfig::new_insecure_only(
-                IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 0x1)),
-                DEFAULT_PORT
-            ),
-            maxcon: MAXIMUM_CONNECTION_LIMIT
-        }
-    );
-}
-
-#[test]
-fn test_config_file_template() {
-    let file = get_toml_from_examples_dir("template.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(
-        cfg,
-        ConfigurationSet::new(
-            false,
-            BGSave::default(),
-            SnapshotConfig::Enabled(SnapshotPref::new(3600, 4, true)),
-            PortConfig::new_secure_only(
-                DEFAULT_IPV4,
-                SslOpts::new(
-                    "/path/to/keyfile.pem".into(),
-                    "/path/to/chain.pem".into(),
-                    2004,
-                    Some("/path/to/cert/passphrase.txt".to_owned())
-                )
-            ),
-            MAXIMUM_CONNECTION_LIMIT
-        )
-    );
-}
-
-#[test]
-fn test_config_file_bad_bgsave_section() {
-    let file = get_toml_from_examples_dir("badcfg2.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file);
-    assert!(cfg.is_err());
-}
-
-#[test]
-fn test_config_file_custom_bgsave() {
-    let file = get_toml_from_examples_dir("withcustombgsave.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(
-        cfg,
-        ConfigurationSet {
-            noart: false,
-            bgsave: BGSave::new(true, 600),
-            snapshot: SnapshotConfig::default(),
-            ports: PortConfig::default(),
-            maxcon: MAXIMUM_CONNECTION_LIMIT
-        }
-    );
-}
-
-#[test]
-fn test_config_file_bgsave_enabled_only() {
-    /*
-     * This test demonstrates a case where the user just said that BGSAVE is enabled.
-     * In that case, we will default to the 120 second duration
-     */
-    let file = get_toml_from_examples_dir("bgsave-justenabled.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(
-        cfg,
-        ConfigurationSet {
-            noart: false,
-            bgsave: BGSave::default(),
-            snapshot: SnapshotConfig::default(),
-            ports: PortConfig::default(),
-            maxcon: MAXIMUM_CONNECTION_LIMIT
-        }
-    )
-}
-
-#[test]
-fn test_config_file_bgsave_every_only() {
-    /*
-     * This test demonstrates a case where the user just gave the value for every
-     * In that case, it means BGSAVE is enabled and set to `every` seconds
-     */
-    let file = get_toml_from_examples_dir("bgsave-justevery.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(
-        cfg,
-        ConfigurationSet {
-            noart: false,
-            bgsave: BGSave::new(true, 600),
-            snapshot: SnapshotConfig::default(),
-            ports: PortConfig::default(),
-            maxcon: MAXIMUM_CONNECTION_LIMIT
-        }
-    )
-}
-
-#[test]
-fn test_config_file_snapshot() {
-    let file = get_toml_from_examples_dir("snapshot.toml".to_owned()).unwrap();
-    let cfg = ConfigurationSet::new_from_toml_str(file).unwrap();
-    assert_eq!(
-        cfg,
-        ConfigurationSet {
-            snapshot: SnapshotConfig::Enabled(SnapshotPref::new(3600, 4, true)),
-            bgsave: BGSave::default(),
-            noart: false,
-            ports: PortConfig::default(),
-            maxcon: MAXIMUM_CONNECTION_LIMIT
-        }
-    );
-}
-
-#[test]
-fn test_cli_args_conflict() {
-    let cfg_layout = load_yaml!("../cli.yml");
-    let cli_args = ["skyd", "--nosave", "-c config.toml"];
-    let matches = App::from_yaml(cfg_layout).get_matches_from(&cli_args);
-    let err = super::get_config_file_or_return_cfg_from_matches(matches).unwrap_err();
-    assert_eq!(err, ConfigError::CfgError(ERR_CONFLICT));
-}
-
-#[test]
-fn test_cli_args_conflict_with_restore_file_okay() {
-    let cfg_layout = load_yaml!("../cli.yml");
-    let cli_args = ["skyd", "--restore", "somedir", "-c", "config.toml"];
-    let matches = App::from_yaml(cfg_layout).get_matches_from(&cli_args);
-    let ret = super::get_config_file_or_return_cfg_from_matches(matches).unwrap_err();
-    // this should only compain about the missing dir but not about conflict
-    assert_eq!(
-        ret,
-        ConfigError::OSError(std::io::Error::from(std::io::ErrorKind::NotFound))
-    );
-}
-
-#[test]
-fn test_cli_args_conflict_with_restore_file_fail() {
-    let cfg_layout = load_yaml!("../cli.yml");
-    let cli_args = [
-        "skyd",
-        "--restore",
-        "somedir",
-        "-c",
-        "config.toml",
-        "--nosave",
-    ];
-    let matches = App::from_yaml(cfg_layout).get_matches_from(&cli_args);
-    let ret = super::get_config_file_or_return_cfg_from_matches(matches).unwrap_err();
-    // this should only compain about the missing dir but not about conflict
-    assert_eq!(ret, ConfigError::CfgError(ERR_CONFLICT));
 }
